Make Notification Drawer Extensible and supports Vue Custom Elements.
