+++ webapp/portlet/.eslintrc.json
@@ -11,6 +11,7 @@
     "activityComposer": true,
     "socialUIProfile": true,
     "QRCode": true,
+    "DOMPurify": true,
     "ExtendedDomPurify": true,
     "Cropper": true,
     "cCometd": true
