+++ webapp/portlet/src/main/webapp/WEB-INF/gatein-resources.xml
@@ -1288,6 +1288,7 @@
   <portlet>
     <name>TopBarNotification</name>
     <module>
+      <load-group>TopBarNotificationGRP</load-group>
       <script>
         <minify>false</minify>
         <path>/js/topBarNotification.bundle.js</path>
@@ -2032,7 +2033,7 @@
     <name>intranetNotification</name>
     <load-group>webNotificationsGRP</load-group>
     <script>
-      <path>/js/notification/IntranetNotification.js</path>
+      <path>/js/webuiNotification/IntranetNotification.js</path>
     </script>
     <depends>
       <module>webNotifications</module>
@@ -2047,7 +2048,7 @@
     <name>webNotifications</name>
     <load-group>webNotificationsGRP</load-group>
     <script>
-      <path>/js/notification/WebNotification.js</path>
+      <path>/js/webuiNotification/WebNotification.js</path>
     </script>
     <depends>
       <module>jquery</module>
