+++ webapp/portlet/src/main/webapp/vue-apps/notification-top-bar/components/NotificationItem.vue
@@ -1,0 +1,141 @@
+<template>
+  <a
+    :id="id"
+    :href="notifLink"
+    class="notifDrawerItem">
+    <dynamic-html-element :child="bodyElement" />
+  </a>
+</template>
+<script>
+const SLIDE_UP = 300;
+const SLIDE_UP_MORE = 600;
+
+export default {
+  props: {
+    id: {
+      type: String,
+      default: null,
+    },
+    notif: {
+      type: Object,
+      default: null,
+    },
+  },
+  data: () => ({
+    notifLink: null,
+  }),
+  computed: {
+    bodyElement() {
+      const template = DOMPurify.sanitize(this.notif.notification || '', {
+        CUSTOM_ELEMENT_HANDLING: {
+          tagNameCheck: () => true,
+          attributeNameCheck: () => true,
+        },
+      });
+      return {template};
+    },
+  },
+  mounted() {
+    this.applyActions();    
+  },
+  methods: {
+    applyActions() {
+      const self = this;
+      $(`#${this.id}`).find('li').each(function () {
+        let dataLink = $(this).find('.contentSmall:first').data('link');
+        if (!dataLink) {
+          dataLink = $(this).find('.media').children('a').attr('href');
+        }
+        if (dataLink.includes('/portal/g/') || dataLink.includes('/portal/rest/')){
+          dataLink.replace(/\/portal\//,`${eXo.env.portal.context}/`);
+        } else {
+          dataLink.replace(/\/portal\/([a-zA-Z0-9_-]+)\//, `${eXo.env.portal.context}/${eXo.env.portal.portalName}/`);
+        }
+        const linkId = dataLink.split(`${eXo.env.portal.context}/`);
+        const dataId = $(this).data('id').toString();
+        const dataDetails = $(this).data('details') && $(this).data('details').toString() ? $(this).data('details').toString() : null;
+        if (linkId != null && linkId.length >1 ) {
+          if (linkId[0].includes('/view_full_activity/')) {
+            const id = linkId[0].split('/view_full_activity/')[1];
+            self.notifLink = `${eXo.env.portal.context}/${eXo.env.portal.portalName}/activity?id=${id}`;
+          } else {
+            self.notifLink = `${eXo.env.portal.context}/${linkId[1]}`;
+          }
+        } else if (dataDetails !== 'notification-details-drawer') {
+          self.notifLink = dataLink.replace(/^\/rest\//,`${eXo.env.portal.context}/rest/`);
+        }
+
+        // ------------- Open details drawer
+        $(this).find('.open-details-drawer').off('click').on('click', function(evt) {
+          evt.preventDefault();
+          evt.stopPropagation();
+          const notificationDetails = $(this).data('notification-details');
+          document.dispatchEvent(new CustomEvent('open-notification-details-drawer', {detail: notificationDetails}));
+          if ($(this).closest('[data-details]').hasClass('unread')) {
+            $(this).closest('[data-details]').removeClass('unread').addClass('read');
+          }
+          Vue.prototype.$notificationService.updateNotification(dataId, 'markAsRead');
+        });
+
+        // ----------------- Mark as read
+        $(this).off('click').on('click', function() {
+          if ($(this).hasClass('unread')) {
+            $(this).removeClass('unread').addClass('read');
+          }
+          Vue.prototype.$notificationService.updateNotification(dataId, 'markAsRead');
+        });
+
+        // ------------- hide notif
+        $(this).find('.remove-item').off('click').on('click', function(evt) {
+          evt.preventDefault();
+          evt.stopPropagation();
+          Vue.prototype.$notificationService.updateNotification(dataId,'hide');
+          $(this).parents('li:first').slideUp(SLIDE_UP);
+        });
+
+        // ------------- Accept request
+        $(this).find('.action-item').off('click').on('click', function(evt) {
+          evt.stopPropagation();
+          evt.preventDefault();
+          let restURl = $(this).data('rest');
+          if (restURl.indexOf('?') >= 0 ) {
+            restURl += '&';
+          } 
+          else {
+            restURl += '?';
+          }
+          restURl += `portal:csrf=${eXo.env.portal.csrfToken}`;
+          if (restURl && restURl.length > 0) {
+            $.ajax(restURl).done(function () {
+              $(document).trigger('exo-invitation-updated');
+            });
+          }
+          Vue.prototype.$notificationService.updateNotification(dataId,'hide');
+          $(this).parents('li:first').slideUp(SLIDE_UP_MORE);
+        });
+
+        // ------------- Refuse request
+        $(this).find('.cancel-item').off('click').on('click', function(evt) {
+          evt.stopPropagation();
+          evt.preventDefault();
+          let restCancelURl = $(this).data('rest');
+          if (restCancelURl.indexOf('?') >= 0 ) {
+            restCancelURl += '&';
+          }
+          else {
+            restCancelURl += '?';
+          }
+          restCancelURl += `portal:csrf=${eXo.env.portal.csrfToken}`;
+          if (restCancelURl && restCancelURl.length > 0) {
+            $.ajax(restCancelURl).done(function () {
+              $(document).trigger('exo-invitation-updated');
+            });
+          }
+          Vue.prototype.$notificationService.updateNotification(dataId,'hide');
+          $(this).parents('li:first').slideUp(SLIDE_UP_MORE);
+        });
+      });
+    },
+  },
+};
+</script>
